+++ tests/actions.rs
@@ -118,7 +118,7 @@ fn post_to_database_ok() {
                                .insert("name", "Babe Ruth")
                                .insert("career_hr", 714)
                                .unwrap();
-    let (rev, doc_id) = client.post_to_database("/baseball", &expected_content).run().unwrap();
+    let (doc_id, rev) = client.post_to_database("/baseball", &expected_content).run().unwrap();
     let doc = client.get_document(("/baseball", doc_id.clone()))
                     .run()
                     .unwrap()
@@ -147,7 +147,7 @@ fn head_document_ok_without_revision() {
                              .insert("name", "Babe Ruth")
                              .insert("career_hr", 714)
                              .unwrap();
-    let (_rev, doc_id) = client.post_to_database("/baseball", &source_content).run().unwrap();
+    let (doc_id, _rev) = client.post_to_database("/baseball", &source_content).run().unwrap();
     let got = client.head_document(("/baseball", doc_id)).run().unwrap();
     assert!(got.is_some());
 }
@@ -160,7 +160,7 @@ fn head_document_ok_fresh_revision() {
                              .insert("name", "Babe Ruth")
                              .insert("career_hr", 714)
                              .unwrap();
-    let (rev, doc_id) = client.post_to_database("/baseball", &source_content).run().unwrap();
+    let (doc_id, rev) = client.post_to_database("/baseball", &source_content).run().unwrap();
     let got = client.head_document(("/baseball", doc_id)).if_none_match(&rev).run().unwrap();
     assert!(got.is_none());
 }
@@ -173,7 +173,7 @@ fn head_document_ok_stale_revision() {
                              .insert("name", "Babe Ruth")
                              .insert("career_hr", 714)
                              .unwrap();
-    let (rev1, doc_id) = client.post_to_database("/baseball", &source_content).run().unwrap();
+    let (doc_id, rev1) = client.post_to_database("/baseball", &source_content).run().unwrap();
     let _rev2 = client.put_document(("/baseball", doc_id.clone()), &source_content)
                       .if_match(&rev1)
                       .run()
@@ -198,7 +198,7 @@ fn get_document_ok_without_revision() {
                                .insert("name", "Babe Ruth")
                                .insert("career_hr", 714)
                                .unwrap();
-    let (rev, doc_id) = client.post_to_database("/baseball", &expected_content).run().unwrap();
+    let (doc_id, rev) = client.post_to_database("/baseball", &expected_content).run().unwrap();
     let got = client.get_document(("/baseball", doc_id.clone()))
                     .run()
                     .unwrap()
@@ -216,7 +216,7 @@ fn get_document_ok_fresh_revision() {
                              .insert("name", "Babe Ruth")
                              .insert("career_hr", 714)
                              .unwrap();
-    let (rev, doc_id) = client.post_to_database("/baseball", &source_content).run().unwrap();
+    let (doc_id, rev) = client.post_to_database("/baseball", &source_content).run().unwrap();
     let got = client.get_document(("/baseball", doc_id.clone()))
                     .if_none_match(&rev)
                     .run()
@@ -232,7 +232,7 @@ fn get_document_ok_stale_revision() {
                                .insert("name", "Babe Ruth")
                                .insert("career_hr", 714)
                                .unwrap();
-    let (rev1, doc_id) = client.post_to_database("/baseball", &expected_content).run().unwrap();
+    let (doc_id, rev1) = client.post_to_database("/baseball", &expected_content).run().unwrap();
     let rev2 = client.put_document(("/baseball", doc_id.clone()), &expected_content)
                      .if_match(&rev1)
                      .run()
@@ -281,7 +281,7 @@ fn put_document_ok_update_document() {
                              .insert("name", "Babe Ruth")
                              .insert("career_hr", 714)
                              .unwrap();
-    let (rev1, doc_id) = client.post_to_database("/baseball", &source_content).run().unwrap();
+    let (doc_id, rev1) = client.post_to_database("/baseball", &source_content).run().unwrap();
     let expected_content = serde_json::builder::ObjectBuilder::new()
                                .insert("name", "Babe Ruth")
                                .insert("career_hr", 714)
@@ -308,7 +308,7 @@ fn put_document_nok_stale_revision() {
                              .insert("name", "Babe Ruth")
                              .insert("career_hr", 714)
                              .unwrap();
-    let (rev1, doc_id) = client.post_to_database("/baseball", &source_content).run().unwrap();
+    let (doc_id, rev1) = client.post_to_database("/baseball", &source_content).run().unwrap();
     let source_content = serde_json::builder::ObjectBuilder::new()
                              .insert("name", "Babe Ruth")
                              .insert("career_hr", 714)
@@ -343,7 +343,7 @@ fn delete_document_ok() {
                              .insert("name", "Babe Ruth")
                              .insert("career_hr", 714)
                              .unwrap();
-    let (rev1, doc_id) = client.post_to_database("/baseball", &source_content).run().unwrap();
+    let (doc_id, rev1) = client.post_to_database("/baseball", &source_content).run().unwrap();
     let rev2 = client.delete_document(("/baseball", doc_id.clone()), &rev1).run().unwrap();
     assert_eq!(rev1.update_number() + 1, rev2.update_number());
     let got = client.head_document(("/baseball", doc_id)).run();
@@ -358,7 +358,7 @@ fn delete_document_nok_stale_revision() {
                              .insert("name", "Babe Ruth")
                              .insert("career_hr", 714)
                              .unwrap();
-    let (rev1, doc_id) = client.post_to_database("/baseball", &source_content).run().unwrap();
+    let (doc_id, rev1) = client.post_to_database("/baseball", &source_content).run().unwrap();
     let _rev2 = client.put_document(("/baseball", doc_id.clone()), &source_content)
                       .if_match(&rev1)
                       .run()
@@ -398,12 +398,12 @@ fn get_view_nonempty_result() {
                              .insert("name", "Babe Ruth")
                              .insert("career_hr", 714)
                              .unwrap();
-    let (_, babe_ruth_id) = client.post_to_database("/baseball", &source_content).run().unwrap();
+    let (babe_ruth_id, _) = client.post_to_database("/baseball", &source_content).run().unwrap();
     let source_content = serde_json::builder::ObjectBuilder::new()
                              .insert("name", "Hank Aaron")
                              .insert("career_hr", 755)
                              .unwrap();
-    let (_, hank_aaron_id) = client.post_to_database("/baseball", &source_content).run().unwrap();
+    let (hank_aaron_id, _) = client.post_to_database("/baseball", &source_content).run().unwrap();
     let design = couchdb::DesignBuilder::new()
                      .build_view("by_career_hr",
                                  "function(doc) { emit(doc.name, doc.career_hr); }",
@@ -438,7 +438,7 @@ fn get_view_with_endkey() {
                              .insert("name", "Babe Ruth")
                              .insert("career_hr", 714)
                              .unwrap();
-    let (_, babe_ruth_id) = client.post_to_database("/baseball", &source_content).run().unwrap();
+    let (babe_ruth_id, _) = client.post_to_database("/baseball", &source_content).run().unwrap();
     let source_content = serde_json::builder::ObjectBuilder::new()
                              .insert("name", "Hank Aaron")
                              .insert("career_hr", 755)
@@ -478,7 +478,7 @@ fn get_view_with_startkey() {
                              .insert("name", "Hank Aaron")
                              .insert("career_hr", 755)
                              .unwrap();
-    let (_, hank_aaron_id) = client.post_to_database("/baseball", &source_content).run().unwrap();
+    let (hank_aaron_id, _) = client.post_to_database("/baseball", &source_content).run().unwrap();
     let design = couchdb::DesignBuilder::new()
                      .build_view("by_career_hr",
                                  "function(doc) { emit(doc.name, doc.career_hr); }",